moved WHA up outside of topics directory
